Unify context.
